Finished adding the camera functionality. Next step -> add the functionality to take a picture and then save it to camera role.
after that try to add pre built filters, etc, etc.

once that is complete. I will add the main functionality of the app. Using maybe AI or a camera engine I will convert the Image to have a digital camera effect. I want the camera to process the image for a long time, and make it have the same effect a digital camera will.